create ArchiveManager from ArchiveDayService refactor ArchiveController with ArchiveManager
